requirements.txt
# Django core framework
Django==5.2.0

# Authentication
django-allauth==0.61.1

# CSS Framework
django-tailwind==3.8.0

# HTTP library for Python
requests==2.32.3

# JSON Web Token implementation
PyJWT==2.10.1

# Cryptographic recipes and primitives 
cryptography==44.0.2    

# Database drivers
psycopg2==2.9.9
sqlparse==0.5.0

# ASGI support
asgiref==3.8.1

# Production server
gunicorn==22.0.0

# Static files handling
whitenoise==6.7.0

# Environment variable management
python-decouple==3.8

# Admin import/export functionality
django-import-export==3.3.1  

# Utilities
Pillow>=10.0.0          # Image processing
pytz==2023.3            # Timezone support
django-resized==1.0.2   # Image resizing for models
psutil==7.0.0           # System resource monitoring